+++ src/test/c_api/config_standard_resnet.json
@@ -1,0 +1,27 @@
+{
+    "model_config_list": [
+        {"config": {
+                "name": "dummy",
+                "base_path": "/ovms/src/test/dummy",
+                "plugin_config": {
+                    "NUM_STREAMS":1,
+                    "AFFINITY": "CORE",
+                    "PERFORMANCE_HINT": "THROUGHPUT",
+                    "PERF_COUNT":"NO"
+                },
+                "shape": "(1, 10)",
+                "nireq": 1}
+        },
+        {"config": {
+                "name": "resnet",
+                "base_path": "/models/resnet-50-tf",
+                "plugin_config": {
+                    "NUM_STREAMS":12,
+                    "AFFINITY": "CORE",
+                    "PERFORMANCE_HINT": "THROUGHPUT",
+                    "PERF_COUNT":"NO"
+                },
+                "nireq": 12}
+        }
+    ]
+}
